# api/api.yml
swagger: "2.0"
info:
  title: CtoFunds REST API
  description: |
    定义CtoFunds服务端的API，REST风格，包括请求响应URI、状态和报文格式
  version: "1.0"
  contact:
    name: 唐睿
    email: "tangrui@ctofunds.com"
host: www.ctofunds.com
schemes:
  - https
  - http
basePath: /api/v1
...
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  token:
    description: 通过登录/login得到的JWT token
    type: apiKey
    name: X-AUTH-TOKEN
    in: header
security:
  - token: []
paths:
...
  /topics/:
    get:
      summary: 查看主题列表
      description: |
        查看所有主题，支持分页和过滤器
      parameters:
        - name: page
          in: path
          description: 当前页码
          required: true
          type: integer
          minimum: 1
        - name: pageSize
          in: path
          description: 每页项数
          required: true
          type: integer
          minimum: 1
        - name: userId
          in: path
          description: 主题发布者ID
          required: false
          type: integer
          format: int64
        - name: title
          in: path
          description: 主题标题
          required: false
          type: string
        - name: category
          in: path
          description: 主题类别
          required: false
          type: string
      tags:
        - topic
      responses:
        200:
          description: 主题列表
          schema:
            properties:
              total:
                description: 主题总数
                type: integer
                minimum: 0
              page:
                description: 当前第几页
                type: integer
                minimum: 1
              pageSize:
                description: 每页最大的主题数
                type: integer
                minimum: 1
              data:
                type: array
                items:
                  $ref: '#/definitions/Topic'
    post:
      summary: 创建主题
      description: |
        根据请求创建主题
      tags:
        - topic
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/Topic'
      responses:
        201:
          description: 成功创建主题
          schema:
            $ref: '#/definitions/Topic'
        400: *400
        401: *401
        403: *403
        500: *500
  /topics/{id}/:
    get:
      summary: 查询主题
      description: |
        得到指定主题的具体信息
      tags:
        - topic
      parameters:
        - &TopicIdPara
          name: id
          in: path
          description: 主题ID
          required: true
          type: integer
          format: int64
      tags:
        - topic
      responses:
        200:
          description: 成功查询主题
          schema:
            $ref: '#/definitions/Topic'
        404: *404
    delete:
      summary: 删除主题
      description: |
        删除指定主题，后端采用懒删除，只是将状态设成已删除，数据将根据一定物理策略删除
      tags:
        - topic
      parameters:
        - *TopicIdPara
      responses:
        200:
          description: 成功删除主题
          schema:
            $ref: '#/definitions/Topic'
        400: *400
        401: *401
        403: *403
        404: *404
        500: *500
    put:
      summary: 更新主题
      description: |
        根据请求中指定要更新的属性对主题进行更新
      tags:
        - topic
      parameters:
        - *TopicIdPara
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/UpdateTopicRequest'
      responses:
        200:
          description: 成功更新主题
          schema:
            $ref: '#/definitions/Topic'
        400: *400
        401: *401
        403: *403
        404: *404
        500: *500
...
definitions:
...
  Topic:
    description:
      表示主题资源
    required:
      - userId
      - title
    properties:
      id:
        description: 主题ID，由后端自动生成，创建主题时不需要
        type: integer
        format: int64
      userId:
        description: 主题发布者
        type: integer
        format: int64
      title:
        description: 主题标题
        type: string
      content:
        description: 主题内容
        type: string
      ctocoins:
        description: 主题悬赏分
        type: string
      confirmed:
        description: 主题是否已有回复被发布者采纳
        type: boolean
      createdAt:
        description: 创建时间
        type: string
        format: date-time
      modifiedAt:
        description: 最近一次修改时间
        type: string
        format: date-time
...